requirements.txt
# DevOps tools
black==22.3.0
click==8.1.3
pytest==7.1.3
pytest-cov==4.0.0
requests==2.26.0

# Rust-based linter
ruff==0.0.284

# Databricks-specific libraries
pandas==1.3.5
python-dotenv==1.0.0
databricks-sql-connector==2.4.0
